+++ database/sql/src/main/scala/cromwell/database/sql/tables/DockerHashStoreEntry.scala
@@ -6,5 +6,5 @@ case class DockerHashStoreEntry
   dockerTag: String,
   dockerHash: String,
   dockerSize: Option[Long],
-  dockerHashStoreEntryId: Option[Int] = None
+  dockerHashStoreEntryId: Option[Long] = None
 )
