+++ database/sql/src/main/scala/cromwell/database/sql/tables/SubWorkflowStoreEntry.scala
@@ -2,11 +2,11 @@ package cromwell.database.sql.tables
 
 case class SubWorkflowStoreEntry
 (
-  rootWorkflowId: Option[Int],
+  rootWorkflowId: Option[Long],
   parentWorkflowExecutionUuid: String,
   callFullyQualifiedName: String,
   callIndex: Int,
   callAttempt: Int,
   subWorkflowExecutionUuid: String,
-  subWorkflowStoreEntryId: Option[Int] = None
+  subWorkflowStoreEntryId: Option[Long] = None
 )
